+++ test/performance/alignment/local_affine_alignment_benchmark.cpp
@@ -30,7 +30,7 @@
 constexpr auto local_affine_cfg = seqan3::align_cfg::method_local |
                                   seqan3::align_cfg::gap{seqan3::gap_scheme{seqan3::gap_score{-1},
                                                                             seqan3::gap_open_score{-10}}} |
-                                  seqan3::align_cfg::scoring{
+                                  seqan3::align_cfg::scoring_scheme{
                                         seqan3::nucleotide_scoring_scheme{seqan3::match_score{4},
                                                                           seqan3::mismatch_score{-5}}
                                   };