+++ test/unit/alignment/pairwise/semi_global_affine_banded_test.cpp
@@ -35,8 +35,8 @@ struct pairwise_semiglobal_affine_banded : public ::testing::Test
         return seqan3::align_cfg::method_global{} |
                seqan3::align_cfg::gap{seqan3::gap_scheme{seqan3::gap_score{-1},
                                                          seqan3::gap_open_score{-10}}} |
-               seqan3::align_cfg::scoring{seqan3::nucleotide_scoring_scheme{seqan3::match_score{4},
-                                                                            seqan3::mismatch_score{-5}}};
+               seqan3::align_cfg::scoring_scheme{seqan3::nucleotide_scoring_scheme{seqan3::match_score{4},
+                                                                                   seqan3::mismatch_score{-5}}};
     }
 };